Update to v2.0.0 (#2) (#3)
* Update to v2.0 template

* Updating to python 3.10

* Removing deprecated op_dict flag

* Updating paths

* Removing deprecated op_dict instruction

* Updating numpy intc data type

* Adding validation label

* Updating gdown to allow dataset donwload

* Adding aikit-tensorflow

* Adding intel packages to introduction

* Updating CFD solutions readability

* Adding intel packages description

* adding intel package links at Solution Technical Overview

* Fix tensorflow sintax

* Fixing typos

* Updating u-net_architecture diagram

* Updating u-net_architecture diagram

* Removing  jupyter notebooks reference at Validated Hardware Details

* Updating diagram to reflect intel packages

* Expanding intel packages names

* Updating workspace name

* Fixing typo

* Updating packages table

* Fixing typos

* Fixing typos

* Fixing typos

* Reorganizing directories creation

* Updating python execution path

* data folder creation no longer needed

* Updating leaf packages

* Updating keras name

* Fixing typo

* Fixing typo

* Updating Intel trademark

* fixing typo

* fixing typo

* fixing typo

* Adding package references at learn more section

* Updating license year to 2024

* Fixing tensorflow typo

* Updating intel packages description

* Updating tpp pakcage reference

* Updating NN description

* Removing INC acronyms

* Updating gdown command

* Removing outdated issue

* Updating copyright year to 2024

* Updating dataset typo

+++ src/utils/image_processing.py
@@ -12,7 +12,7 @@ from PIL import Image, ImageOps
 def add_padding(img, pad_l, pad_t, pad_r):
     height, width = img.shape
     #Adding padding to the left side.
-    pad_left = np.zeros((height, pad_l), dtype=np.int)
+    pad_left = np.zeros((height, pad_l), dtype=np.intc)
     img = np.concatenate((pad_left, img), axis=1)
 
     #Adding padding to the top.
